+++ src/CommandingOfficers/CommanderLibrary.java
@@ -3,6 +3,8 @@ package CommandingOfficers;
 import java.util.ArrayList;
 
 import CommandingOfficers.AW4.BrennerWolves.*;
+import CommandingOfficers.DefendPeace.CyanOcean.Ave;
+import CommandingOfficers.DefendPeace.CyanOcean.Patch;
 import CommandingOfficers.DefendPeace.misc.Bear_Bull;
 import CommandingOfficers.DefendPeace.misc.DocLight;
 import CommandingOfficers.DefendPeace.misc.Meridian;
