+++ src/CommandingOfficers/DefendPeace/CyanOcean/Ave.java
@@ -1,4 +1,4 @@
-package CommandingOfficers;
+package CommandingOfficers.DefendPeace.CyanOcean;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -8,6 +8,10 @@ import java.util.PriorityQueue;
 import java.util.Set;
 import java.util.Stack;
 
+import CommandingOfficers.Commander;
+import CommandingOfficers.CommanderAbility;
+import CommandingOfficers.CommanderInfo;
+import CommandingOfficers.CommanderInfo.InfoPage;
 import Engine.GameInstance;
 import Engine.GameScenario;
 import Engine.Utils;
