+++ src/CommandingOfficers/DefendPeace/CyanOcean/Patch.java
@@ -1,7 +1,12 @@
-package CommandingOfficers;
+package CommandingOfficers.DefendPeace.CyanOcean;
 
 import java.util.ArrayList;
 
+import CommandingOfficers.Commander;
+import CommandingOfficers.CommanderAbility;
+import CommandingOfficers.CommanderInfo;
+import CommandingOfficers.CommanderAbility.CostBasis;
+import CommandingOfficers.CommanderInfo.InfoPage;
 import Engine.GameInstance;
 import Engine.GameScenario;
 import Engine.GameEvents.GameEventListener;
