+++ src/Test/TestAIBehavior.java
@@ -5,8 +5,8 @@ import AI.AIMaker;
 import AI.Muriel;
 import AI.WallyAI;
 import CommandingOfficers.Commander;
-import CommandingOfficers.Patch;
 import CommandingOfficers.Strong;
+import CommandingOfficers.DefendPeace.CyanOcean.Patch;
 import Engine.Army;
 import Engine.GameAction;
 import Engine.GameInstance;
