+++ src/Test/TestCommanderBrenner.java
@@ -4,9 +4,9 @@ import java.util.ArrayList;
 
 import CommandingOfficers.Commander;
 import CommandingOfficers.DeployableCommander;
-import CommandingOfficers.Patch;
 import CommandingOfficers.AW4.RuinedCommander;
 import CommandingOfficers.AW4.RuinedCommander.UseAbilityFactory;
+import CommandingOfficers.DefendPeace.CyanOcean.Patch;
 import CommandingOfficers.AW4.BrennerWolves.BrennerDoR;
 import Engine.Army;
 import Engine.GameActionSet;
