+++ src/Test/TestGameEvent.java
@@ -1,7 +1,7 @@
 package Test;
 
 import CommandingOfficers.Commander;
-import CommandingOfficers.Patch;
+import CommandingOfficers.DefendPeace.CyanOcean.Patch;
 import Engine.Army;
 import Engine.GameAction;
 import Engine.GameInstance;
